+++ srcs/back/user_mgmt/login/templates/signup_form.html
@@ -1,6 +1,6 @@
 <div id="login-container">
 	<div class="container">
-		<h1>Sign Up</h1>
+		<h1>{{sign_up}}</h1>
 		<form id="signup-form" method="post" action="http://localhost:8000/api/register/">
 			<div style="display: flex; align-items: center; justify-content: space-between;">
 				<label for="username">{{username}}:</label>
@@ -22,7 +22,9 @@
 				<label for="email">{{email}}:</label>
 				<input type="text" id="email" name="email">
 			</div>
-			<button id="signup-submit" type="submit" value="submit">Submit</button>
+			<button id="signup-submit" type="submit" value="submit">{{submit}}</button>
++                       <p id="login" class="mt-2">{{have_account}}<button data-route="/login" 
++                               id="login-link" class="btn btn-link p-0 text-decoration-underline" type="button">{{click_here}}</button></p> 
 		</form>
 	</div>
 </div>
