+++ srcs/back/user_mgmt/login/urls.py
@@ -7,11 +7,11 @@ urlpatterns = [
 	# path("", views.home, name="home"),
 	# path("login/", views.login_view, name="login"),
 	# path("logout/", views.logout_view, name="logout"),
-
 	path("api/login/", views.login_view, name="login_api"),
-	path("api/login-form/", views.login_form_api, name="login_form_api"),
 	path("api/2fa-login/", views.verify_2fa_login_form, name="login_2fa_form"),
 	path("api/2fa-login/verify/", views.verify_login_2fa, name="verify_login_2fa"),
+	path("api/login-form/", views.login_form, name="login_form"),
+	path("api/signup-form/", views.signup_form, name="signup_form"),
 	path("api/register/", views.register_user, name='register_user'),
 	path("api/token/refresh/", TokenRefreshView.as_view(), name='token_refresh'),
 	path("api/login-intra/", auth42.login_intra, name="login_intra"),
