+++ srcs/back/user_mgmt/login/views.py
@@ -21,6 +21,7 @@ from user_profile.models import Profile
 from .twoFA import TwoFA  
 import base64
 from datetime import datetime, timedelta
+from user_mgmt.utils.translations import add_language_context
 
 def generate_jwt_tokens(user):
     refresh = RefreshToken.for_user(user)
@@ -102,15 +103,42 @@ def verify_login_2fa(request):
     except User.DoesNotExist:
         return Response({'success': False, 'message': 'User does not exist.'}, status=404)
 
+# @api_view(['GET'])
+# @permission_classes([AllowAny])
+# def login_form_api(request):
+#     if request.method == "GET":
+#         print("Login form API called")
+#         form_html = render_to_string('login.html')
+#         return JsonResponse({'form_html': form_html}, content_type="application/json")
+#     else:
+#         return JsonResponse({'error': 'Invalid request method'}, status=405)
+
 @api_view(['GET'])
 @permission_classes([AllowAny])
-def login_form_api(request):
+def login_form(request):
     if request.method == "GET":
         print("Login form API called")
-        form_html = render_to_string('login.html')
+        context = {}
+        add_language_context(request, context)
+        form_html = render_to_string('login_form.html', context)
         return JsonResponse({'form_html': form_html}, content_type="application/json")
     else:
         return JsonResponse({'error': 'Invalid request method'}, status=405)
+
+@api_view(['GET'])
+@permission_classes([AllowAny])
+def signup_form(request):
+    print("Signup method called")
+    if request.method == "GET":
+        print("Signup form API called")
+        context = {}
+        add_language_context(request, context)
+        form_html = render_to_string('signup_form.html', context)
+        return JsonResponse({'form_html': form_html}, content_type="application/json")
+    else:
+        return JsonResponse({'error': 'Invalid request method'}, status=405)
+
+
 
 @api_view(['GET'])
 @permission_classes([AllowAny])
