+++ srcs/back/user_mgmt/user_profile/models.py
@@ -8,7 +8,7 @@ class Profile(models.Model):
     external_photo_url = models.URLField(blank=True, null=True)
     friends = models.ManyToManyField('self', symmetrical=False, related_name='friend_of', blank=True)
     online_status = models.BooleanField(default=False)
-    language = models.CharField(max_length=10, default="en")
+    language = models.CharField(max_length=10, default="")
     two_fa = models.BooleanField(default=False)
     totp_secret = models.CharField(max_length=32, blank=True, null=True)
